+++ include/envoy/tracing/http_tracer.h
@@ -40,6 +40,40 @@ struct Decision {
 };
 
 /**
+ * The context for the custom tag to obtain the tag value.
+ */
+struct CustomTagContext {
+  const Http::HeaderMap* request_headers;
+  const StreamInfo::StreamInfo& stream_info;
+};
+
+class Span;
+using SpanPtr = std::unique_ptr<Span>;
+
+/**
+ * Tracing custom tag, with tag name and how it would be applied to the span.
+ */
+class CustomTag {
+public:
+  virtual ~CustomTag() = default;
+
+  /**
+   * @return the tag name view.
+   */
+  virtual absl::string_view tag() const PURE;
+
+  /**
+   * The way how to apply the custom tag to the span,
+   * generally obtain the tag value from the context and attached it to the span.
+   * @param span the active span.
+   * @param ctx the custom tag context.
+   */
+  virtual void apply(Span& span, const CustomTagContext& ctx) const PURE;
+};
+
+using CustomTagPtr = std::shared_ptr<CustomTag>;
+
+/**
  * Tracing configuration, it carries additional data needed to populate the span.
  */
 class Config {
@@ -52,9 +86,9 @@ public:
   virtual OperationName operationName() const PURE;
 
   /**
-   * @return list of headers to populate tags on the active span.
+   * @return custom tags to be attached to the active span.
    */
-  virtual const std::vector<Http::LowerCaseString>& requestHeadersForTags() const PURE;
+  virtual const std::vector<CustomTagPtr>& customTags() const PURE;
 
   /**
    * @return true if spans should be annotated with more detailed information.
@@ -66,9 +100,6 @@ public:
    */
   virtual uint32_t maxPathTagLength() const PURE;
 };
-
-class Span;
-using SpanPtr = std::unique_ptr<Span>;
 
 /**
  * Basic abstraction for span.