+++ source/common/router/config_impl.h
@@ -338,10 +338,13 @@ public:
   // Tracing::getOverallSampling
   const envoy::type::FractionalPercent& getOverallSampling() const override;
 
+  const std::vector<Tracing::CustomTagPtr>& getCustomTags() const override;
+
 private:
   envoy::type::FractionalPercent client_sampling_;
   envoy::type::FractionalPercent random_sampling_;
   envoy::type::FractionalPercent overall_sampling_;
+  std::vector<Tracing::CustomTagPtr> custom_tags_;
 };
 
 /**