+++ source/common/tracing/http_tracer_impl.h
@@ -2,10 +2,12 @@
 
 #include <string>
 
+#include "envoy/api/v2/core/base.pb.h"
 #include "envoy/local_info/local_info.h"
 #include "envoy/runtime/runtime.h"
 #include "envoy/thread_local/thread_local.h"
 #include "envoy/tracing/http_tracer.h"
+#include "envoy/type/tracing.pb.h"
 #include "envoy/upstream/cluster_manager.h"
 
 #include "common/http/header_map_impl.h"
@@ -118,6 +120,12 @@ public:
                                    const StreamInfo::StreamInfo& stream_info,
                                    const Config& tracing_config);
 
+  /**
+   * Create a custom tag according to the configuration.
+   * @param tag a tracing custom tag configuration.
+   */
+  static CustomTagPtr createCustomTag(const envoy::type::TracingCustomTag& tag);
+
 private:
   static void setCommonTags(Span& span, const Http::HeaderMap* response_headers,
                             const Http::HeaderMap* response_trailers,
@@ -132,14 +140,12 @@ class EgressConfigImpl : public Config {
 public:
   // Tracing::Config
   Tracing::OperationName operationName() const override { return Tracing::OperationName::Egress; }
-  const std::vector<Http::LowerCaseString>& requestHeadersForTags() const override {
-    return request_headers_for_tags_;
-  }
+  const std::vector<CustomTagPtr>& customTags() const override { return custom_tags_; }
   bool verbose() const override { return false; }
   uint32_t maxPathTagLength() const override { return Tracing::DefaultMaxPathTagLength; }
 
 private:
-  const std::vector<Http::LowerCaseString> request_headers_for_tags_{};
+  const std::vector<CustomTagPtr> custom_tags_{};
 };
 
 using EgressConfig = ConstSingleton<EgressConfigImpl>;
@@ -184,6 +190,130 @@ public:
 private:
   DriverPtr driver_;
   const LocalInfo::LocalInfo& local_info_;
+};
+
+class GeneralCustomTag : public CustomTag {
+public:
+  explicit GeneralCustomTag(const std::string& tag) : tag_(tag) {}
+  absl::string_view tag() const override { return tag_; }
+  void apply(Span& span, const CustomTagContext& ctx) const override;
+
+  virtual absl::string_view value(const CustomTagContext& ctx) const PURE;
+  virtual std::string toString() const PURE;
+
+protected:
+  std::string tag_;
+};
+
+class LiteralCustomTag : public GeneralCustomTag {
+public:
+  LiteralCustomTag(const std::string& tag, const envoy::type::TracingCustomTag::Literal& literal)
+      : GeneralCustomTag(tag), value_(literal.value()) {}
+  LiteralCustomTag(const std::string& tag, const std::string& value)
+      : GeneralCustomTag(tag), value_(value) {}
+  absl::string_view value(const CustomTagContext&) const override { return value_; }
+  std::string toString() const override { return fmt::format("LITERAL|{}|{}", tag_, value_); }
+
+private:
+  std::string value_;
+};
+
+class EnvironmentCustomTag : public GeneralCustomTag {
+public:
+  EnvironmentCustomTag(const std::string& tag,
+                       const envoy::type::TracingCustomTag::Environment& environment);
+  EnvironmentCustomTag(const std::string& tag, const std::string& name);
+  EnvironmentCustomTag(const std::string& tag, const std::string& name,
+                       const std::string& default_value);
+  absl::string_view value(const CustomTagContext&) const override { return final_value_; }
+  std::string toString() const override {
+    return fmt::format("ENVIRONMENT|{}|{}|{}", tag_, name_, default_value_);
+  }
+
+private:
+  std::string name_;
+  std::string default_value_;
+  std::string final_value_;
+};
+
+class RequestHeaderCustomTag : public GeneralCustomTag {
+public:
+  RequestHeaderCustomTag(const std::string& tag,
+                         const envoy::type::TracingCustomTag::Header& request_header);
+  RequestHeaderCustomTag(const std::string& tag, const std::string& name);
+  RequestHeaderCustomTag(const std::string& tag, const std::string& name,
+                         const std::string& default_value);
+  absl::string_view value(const CustomTagContext& ctx) const override;
+  std::string toString() const override {
+    return fmt::format("REQUEST_HEADER|{}|{}|{}", tag_, name_.get(), default_value_);
+  }
+
+private:
+  Http::LowerCaseString name_;
+  std::string default_value_;
+};
+
+class MetadataCustomTag : public GeneralCustomTag {
+public:
+  MetadataCustomTag(const std::string& tag,
+                    const envoy::type::TracingCustomTag::Metadata& metadata);
+  MetadataCustomTag(const std::string& tag, const std::string& filter_namespace,
+                    const std::string& raw_path);
+  MetadataCustomTag(const std::string& tag, const std::string& filter_namespace,
+                    const std::string& raw_path, const std::string& separator,
+                    const std::string& default_value);
+  void apply(Span& span, const CustomTagContext& ctx) const override;
+  absl::string_view value(const CustomTagContext&) const override { return default_value_; }
+  virtual const envoy::api::v2::core::Metadata* metadata(const CustomTagContext& ctx) const PURE;
+
+protected:
+  std::string filter_namespace_;
+  std::string raw_path_;
+  std::string separator_;
+  std::string default_value_;
+
+  static const std::string DefaultSeparator;
+
+private:
+  std::vector<std::string> path_;
+};
+
+class RequestMetadataCustomTag : public MetadataCustomTag {
+public:
+  RequestMetadataCustomTag(const std::string& tag,
+                           const envoy::type::TracingCustomTag::Metadata& request_metadata)
+      : MetadataCustomTag(tag, request_metadata) {}
+  RequestMetadataCustomTag(const std::string& tag, const std::string& filter_namespace,
+                           const std::string& raw_path)
+      : MetadataCustomTag(tag, filter_namespace, raw_path) {}
+  RequestMetadataCustomTag(const std::string& tag, const std::string& filter_namespace,
+                           const std::string& raw_path, const std::string& separator,
+                           const std::string& default_value)
+      : MetadataCustomTag(tag, filter_namespace, raw_path, separator, default_value) {}
+  std::string toString() const override {
+    return fmt::format("REQUEST_METADATA|{}|{}|{}|{}", tag_, filter_namespace_, raw_path_,
+                       default_value_);
+  }
+  const envoy::api::v2::core::Metadata* metadata(const CustomTagContext& ctx) const override;
+};
+
+class RouteMetadataCustomTag : public MetadataCustomTag {
+public:
+  RouteMetadataCustomTag(const std::string& tag,
+                         const envoy::type::TracingCustomTag::Metadata& route_metadata)
+      : MetadataCustomTag(tag, route_metadata) {}
+  RouteMetadataCustomTag(const std::string& tag, const std::string& filter_namespace,
+                         const std::string& raw_path)
+      : MetadataCustomTag(tag, filter_namespace, raw_path) {}
+  RouteMetadataCustomTag(const std::string& tag, const std::string& filter_namespace,
+                         const std::string& raw_path, const std::string& separator,
+                         const std::string& default_value)
+      : MetadataCustomTag(tag, filter_namespace, raw_path, separator, default_value) {}
+  std::string toString() const override {
+    return fmt::format("ROUTE_METADATA|{}|{}|{}|{}", tag_, filter_namespace_, raw_path_,
+                       default_value_);
+  }
+  const envoy::api::v2::core::Metadata* metadata(const CustomTagContext& ctx) const override;
 };
 
 } // namespace Tracing