+++ source/extensions/filters/network/http_connection_manager/config.cc
@@ -24,6 +24,7 @@
 #include "common/protobuf/utility.h"
 #include "common/router/rds_impl.h"
 #include "common/router/scoped_rds.h"
+#include "common/tracing/http_tracer_impl.h"
 
 namespace Envoy {
 namespace Extensions {
@@ -243,7 +244,6 @@ HttpConnectionManagerConfig::HttpConnectionManagerConfig(
     const auto& tracing_config = config.tracing();
 
     Tracing::OperationName tracing_operation_name;
-    std::vector<Http::LowerCaseString> request_headers_for_tags;
 
     // Listener level traffic direction overrides the operation name
     switch (context.direction()) {
@@ -272,8 +272,15 @@ HttpConnectionManagerConfig::HttpConnectionManagerConfig(
       NOT_REACHED_GCOVR_EXCL_LINE;
     }
 
+    std::vector<Tracing::CustomTagPtr> custom_tags;
     for (const std::string& header : tracing_config.request_headers_for_tags()) {
-      request_headers_for_tags.push_back(Http::LowerCaseString(header));
+      custom_tags.push_back(std::make_shared<Tracing::RequestHeaderCustomTag>(header, header));
+    }
+    for (const auto& tag : tracing_config.custom_tags()) {
+      Tracing::CustomTagPtr ptr = Tracing::HttpTracerUtility::createCustomTag(tag);
+      if (ptr) {
+        custom_tags.push_back(ptr);
+      }
     }
 
     envoy::type::FractionalPercent client_sampling;
@@ -295,8 +302,8 @@ HttpConnectionManagerConfig::HttpConnectionManagerConfig(
 
     tracing_config_ =
         std::make_unique<Http::TracingConnectionManagerConfig>(Http::TracingConnectionManagerConfig{
-            tracing_operation_name, request_headers_for_tags, client_sampling, random_sampling,
-            overall_sampling, tracing_config.verbose(), max_path_tag_length});
+            tracing_operation_name, custom_tags, client_sampling, random_sampling, overall_sampling,
+            tracing_config.verbose(), max_path_tag_length});
   }
 
   for (const auto& access_log : config.access_log()) {