+++ test/mocks/router/mocks.h
@@ -352,6 +352,7 @@ public:
   MOCK_CONST_METHOD0(getClientSampling, const envoy::type::FractionalPercent&());
   MOCK_CONST_METHOD0(getRandomSampling, const envoy::type::FractionalPercent&());
   MOCK_CONST_METHOD0(getOverallSampling, const envoy::type::FractionalPercent&());
+  MOCK_CONST_METHOD0(getCustomTags, const std::vector<Tracing::CustomTagPtr>&());
 };
 
 class MockRoute : public Route {